+++ service/k8sclient/k8sclient.go
@@ -16,33 +16,49 @@ limitations under the License.
 package k8sclient
 
 import (
+	"errors"
+	"fmt"
 	"os"
-	"path/filepath"
 
+	"github.com/mitchellh/go-homedir"
 	"k8s.io/client-go/kubernetes"
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
 	"k8s.io/client-go/tools/clientcmd"
-	"k8s.io/client-go/util/homedir"
 
 	log "github.com/sirupsen/logrus"
 )
 
-func New(kubeContext, kubeConfig string) (*kubernetes.Clientset, error) {
-	// Set the kubeconfig to the default location if the path wasn't provided.
-	if kubeConfig == "" {
-		// Check if KUBECONFIG environment variable is set
-		if envKubeConfig := os.Getenv("KUBECONFIG"); envKubeConfig != "" {
-			kubeConfig = envKubeConfig
-		} else if home := homedir.HomeDir(); home != "" {
-			kubeConfig = filepath.Join(home, ".kube", "config")
+// findKubeConfig discovers the current kubeconfig
+func findKubeConfig() (string, error) {
+	// Check if KUBECONFIG environment variable is set
+	if envKubeConfig := os.Getenv("KUBECONFIG"); envKubeConfig != "" {
+		return envKubeConfig, nil
+	}
+
+	path, err := homedir.Expand("~/.kube/config")
+	if err != nil {
+		return "", err
+	}
+
+	return path, nil
+}
+
+func New(kubeContext, kubeConfigPath string) (*kubernetes.Clientset, error) {
+	var err error
+
+	// Discover the kubeconfig if an explicit path wasn't provided
+	if kubeConfigPath == "" {
+		kubeConfigPath, err = findKubeConfig()
+		if err != nil {
+			return nil, err
 		}
 	}
 
 	// Inform the user which kubeconfig file is being used
-	log.Infof("using kubeconfig: %s", kubeConfig)
+	log.Infof("using kubeconfig: %s", kubeConfigPath)
 
 	// Set defaults for creating a new ClientConfig.
-	loadingRules := &clientcmd.ClientConfigLoadingRules{ExplicitPath: kubeConfig}
+	loadingRules := &clientcmd.ClientConfigLoadingRules{ExplicitPath: kubeConfigPath}
 	configOverrides := &clientcmd.ConfigOverrides{}
 
 	// Set the context if it was provided.
@@ -63,4 +79,37 @@ func New(kubeContext, kubeConfig string) (*kubernetes.Clientset, error) {
 	}
 
 	return k8sClientSet, err
+}
+
+// GetNamespace returns the current namespace from a Kubeconfig
+func GetNamespace(kubeConfigPath, kubeContext string) (string, error) {
+	var err error
+	var context string
+
+	// Discover the kubeconfig if an explicit path wasn't provided
+	if kubeConfigPath == "" {
+		kubeConfigPath, err = findKubeConfig()
+		if err != nil {
+			return "", err
+		}
+	}
+
+	kubeConfig, err := clientcmd.LoadFromFile(kubeConfigPath)
+	if err != nil {
+		return "", err
+	}
+
+	// Use the context if it was provided, otherwise use the current context
+	if kubeContext != "" {
+		context = kubeContext
+	} else {
+		context = kubeConfig.CurrentContext
+	}
+
+	namespace := kubeConfig.Contexts[context].Namespace
+	if namespace == "" {
+		return "", errors.New(fmt.Sprintf("No current context found in kubeconfig: %s", kubeConfigPath))
+	}
+
+	return namespace, nil
 }